Move field element to it's own module
