+++ py/src/merkle_tree/proof.rs
@@ -1,9 +1,7 @@
 use pyo3::prelude::*;
 
-#[pyclass(name="Proof")]
+#[pyclass(name = "Proof")]
 pub struct PyProof {
     pub merkle_path: Vec<FieldElement<F>>,
     pub hasher: H,
 }
-
-#[pymodule]
